perf: compose bar data table once and filter them for each facet
